requirements.txt
starlette
flask
flask_cors
werkzeug
object-detection-0.1==0.1
poppler-utils
numpy
pdf2image
pillow==8.0.1
python-multipart==0.0.5
pytest==6.1.2